Added a protocol to apply BQSR.
GATK4 no longer supports on-the-fly BQSRing during variant calling, so BQSR needs to be applied in advance. This cannot be done in the variant calling script, because these are done in batches, so this would apply BQSR to the entire sample BAM for each BAM. This also cannot be done after calculating BQSR, because BQSR is currently being calculated before marking duplicates. Thus, applying BQSR needs to be done after calculating BQSR (obviously), and after marking duplicates, but before variant calling.